RasPi/NeoPixel/requirements.txt
adafruit-circuitpython-neopixel
rpi-ws281x
RPi.GPIO
